Merged PR 2: Build and Test pipeline
This PR adds an azure pipeline yaml configuration that builds and tests the main scheme of the project. This will run all Unit and UI tests in the project. It also lowers the project's minimum deployment version to be compatible with the simulators available on Azure Cloud Hosted Agents.

Related work items: #8, #9